lint: the record size is already int64, and the plan cap is a max
golangci-lint v2.12.2's unconvert found the two int64(8 + ...) casts on
expressions that are int64 already: ir.TableFixedTypeBytes returns int64,
so 8 + body and 8 + ir.TableFixedTypeBytes(st) need no conversion.

modernize v0.23.0 never ran in CI because golangci-lint failed ahead of
it; locally it names the plan-cap floor, so that if is a max.

+++ internal/codegen/rusttable/fixedlineage.go
@@ -50,7 +50,7 @@ func FixedLineageOf(u *ir.Unit, name string) (FixedLineageEntry, bool) {
 		return FixedLineageEntry{
 			Wire:   ir.TableFixedLayoutHash(layout, st),
 			Layout: layout,
-			Record: int64(8 + ir.TableFixedTypeBytes(st)),
+			Record: 8 + ir.TableFixedTypeBytes(st),
 		}, true
 	}
 	return FixedLineageEntry{}, false
